crm/cron_jobs/requirements.txt
django-crontab==0.7.1
graphene-django==3.2.0
django-filter==23.5
gql==3.5.0
aiohttp==3.9.5